Extract shared Harbor benchmark runner
